Split viz into files, and make `viz_edge_network` a standalone function in the viz module.
