+++ fvdb/viz/__init__.py
@@ -1,10 +1,13 @@
 # Copyright Contributors to the OpenVDB Project
 # SPDX-License-Identifier: Apache-2.0
 #
+from ._utils import grid_edge_network, gridbatch_edge_network
 from .viewer import CameraView, GaussianSplat3dView, Viewer
 
 __all__ = [
     "Viewer",
     "GaussianSplat3dView",
     "CameraView",
+    "grid_edge_network",
+    "gridbatch_edge_network",
 ]
