+++ fvdb/viz/viewer.py
@@ -1,8 +1,6 @@
 # Copyright Contributors to the OpenVDB Project
 # SPDX-License-Identifier: Apache-2.0
 #
-from typing import Any
-
 import numpy as np
 import torch
 
@@ -12,240 +10,8 @@ from .._Cpp import GaussianSplat3dView as GaussianSplat3dViewCpp
 from .._Cpp import Viewer as ViewerCpp
 from ..gaussian_splatting import GaussianSplat3d
 from ..types import NumericMaxRank1, to_Vec3f
-
-
-class GaussianSplat3dView:
-    """
-    A view for a `GaussianSplat3d` in a `Viewer` with parameters to adjust how the `GaussianSplat3d` is rendered.
-    """
-
-    __PRIVATE__ = object()
-
-    def __init__(self, view: GaussianSplat3dViewCpp, _private: Any = None):
-        """
-        Create a new `GaussianSplat3dView` from a C++ implementation. This constructor is private and should not be called directly.
-        Use `Viewer.register_gaussian_splat3d_view()` instead.
-
-        Args:
-            view (GaussianSplat3dViewCpp): The C++ implementation of the Gaussian splat 3D view.
-            _private (Any): A private object to prevent direct construction. Must be `GaussianSplat3dView.__PRIVATE__`.
-        """
-        self._view = view
-        if _private is not self.__PRIVATE__:
-            raise ValueError("GaussianSplat3dView constructor is private. Use Viewer.add_gaussian_splat3d() instead.")
-
-    @property
-    def tile_size(self) -> int:
-        """
-        Set the 2D tile size to use when rendering splats. Larger tiles can improve performance, but may
-        exhaust shared memory usage on the GPU. In general, tile sizes of 8, 16, or 32 are recommended.
-
-        Returns:
-            int: The current tile size.
-        """
-        return self._view.tile_size
-
-    @tile_size.setter
-    def tile_size(self, tile_size: int):
-        """
-        Set the 2D tile size to use when rendering splats. Larger tiles can improve performance, but may
-        exhaust shared memory usage on the GPU. In general, tile sizes of 8, 16, or 32 are recommended.
-
-        Args:
-            tile_size (int): The tile size to set.
-        """
-        if tile_size < 1:
-            raise ValueError(f"Tile size must be a positive integer, got {tile_size}")
-        self._view.tile_size = tile_size
-
-    @property
-    def min_radius_2d(self) -> float:
-        """
-        Get the minimum radius in pixels below which splats will not be rendered.
-
-        Returns:
-            float: The minimum radius in pixels.
-        """
-        return self._view.min_radius_2d
-
-    @min_radius_2d.setter
-    def min_radius_2d(self, radius: float):
-        """
-        Set the minimum radius in pixels below which splats will not be rendered.
-
-        Args:
-            radius (float): The minimum radius in pixels.
-        """
-        if radius < 0.0:
-            raise ValueError(f"Minimum radius must be non-negative, got {radius}")
-        self._view.min_radius_2d = radius
-
-    @property
-    def sh_degree_to_use(self) -> int:
-        """
-        Get the degree of spherical harmonics to use when rendering colors.
-
-        Returns:
-            int: The degree of spherical harmonics to use.
-        """
-        return self._view.sh_degree_to_use
-
-    @sh_degree_to_use.setter
-    def sh_degree_to_use(self, degree: int):
-        """
-        Sets the degree of spherical harmonics to use when rendering colors. If -1, the maximum
-        degree supported by the Gaussian splat 3D scene is used.
-
-        Args:
-            degree (int): The degree of spherical harmonics to use.
-        """
-        self._view.sh_degree_to_use = degree
-
-    @property
-    def near(self) -> float:
-        """
-        Gets the near clipping plane distance for rendering. Splats closer to the camera than this distance
-        will not be rendered.
-
-        Returns:
-            float: The near clipping plane distance.
-        """
-        return self._view.near
-
-    @near.setter
-    def near(self, near: float):
-        """
-        Sets the near clipping plane distance for rendering. Splats closer to the camera than this distance
-        will not be rendered.
-
-        Args:
-            near (float): The near clipping plane distance.
-        """
-        self._view.near = near
-
-    @property
-    def far(self) -> float:
-        """
-        Get the far clipping plane distance for rendering. Splats farther from the camera than this distance
-        will not be rendered.
-
-        Returns:
-            float: The far clipping plane distance.
-        """
-        return self._view.far
-
-    @far.setter
-    def far(self, far: float):
-        """
-        Sets the far clipping plane distance for rendering. Splats farther from the camera than this distance
-        will not be rendered.
-
-        Args:
-            far (float): The far clipping plane distance.
-        """
-        self._view.far = far
-
-
-class CameraView:
-    """
-    A view for a set of camera frusta and axes in a `Viewer` with parameters to adjust how the cameras are rendered.
-
-    Each camera is represented by its camera-to-world and projection matrices, and drawn as a wireframe frustum
-    with orthogonal axes at the camera origin.
-    """
-
-    __PRIVATE__ = object()
-
-    def __init__(self, view: CameraViewCpp, _private: Any = None):
-        """
-        Create a new `CameraView` from a C++ implementation. This constructor is private and should not be called directly.
-        Use `Viewer.add_camera_view()` instead.
-        """
-        self._view = view
-        if _private is not self.__PRIVATE__:
-            raise ValueError("CameraView constructor is private. Use Viewer.add_camera_view() instead.")
-
-    @property
-    def visible(self) -> bool:
-        """
-        Get whether the camera frusta and axes are shown in the viewer.
-
-        Returns:
-            bool: True if the camera frusta and axes are visible, False otherwise.
-        """
-        return self._view.visible
-
-    @visible.setter
-    def visible(self, visible: bool):
-        """
-        Set whether the camera frusta and axes are shown in the viewer.
-        Args:
-            visible (bool): True to show the camera frusta and axes, False to hide them
-        """
-        self._view.visible = visible
-
-    @property
-    def axis_length(self) -> float:
-        """
-        Get the length of the axes drawn at each camera origin in world units.
-
-        Returns:
-            float: The length of the axes.
-        """
-        return self._view.axis_length
-
-    @axis_length.setter
-    def axis_length(self, length: float):
-        """
-        Set the length of the axes drawn at each camera origin in world units.
-        Args:
-            length (float): The length of the axes.
-        """
-        self._view.axis_length = length
-
-    @property
-    def axis_thickness(self) -> float:
-        """
-        Get the thickness of the axes drawn at each camera origin in world units.
-
-        Returns:
-            float: The thickness of the axes.
-        """
-        return self._view.axis_thickness
-
-    @axis_thickness.setter
-    def axis_thickness(self, thickness: float):
-        """
-        Set the thickness of the axes drawn at each camera origin in world units.
-
-        Args:
-            thickness (float): The thickness of the axes.
-        """
-        self._view.axis_thickness = thickness
-
-    @property
-    def axis_scale(self) -> float:
-        return self._view.axis_scale
-
-    @axis_scale.setter
-    def axis_scale(self, scale: float):
-        self._view.axis_scale = scale
-
-    @property
-    def frustum_line_width(self) -> float:
-        return self._view.frustum_line_width
-
-    @frustum_line_width.setter
-    def frustum_line_width(self, width: float):
-        self._view.frustum_line_width = width
-
-    @property
-    def frustum_scale(self) -> float:
-        return self._view.frustum_scale
-
-    @frustum_scale.setter
-    def frustum_scale(self, scale: float):
-        self._view.frustum_scale = scale
+from ._camera_view import CameraView
+from ._gaussian_splat_3d_view import GaussianSplat3dView
 
 
 class Viewer:
